requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2
django-admin-honeypot==1.1.0
django-admin-thumbnails==0.2.6
django-appconf==1.0.5
django-bootstrap-v5==1.0.11
django-ckeditor==6.4.2
django-imagekit==4.1.0
django-import-export==2.8.0
django-js-asset==2.0.0
django-multiselectfield==0.1.12
django-session-timeout==0.1.0
et-xmlfile==1.1.0
gunicorn==20.1.0
humanize==4.1.0
idna==3.3
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.10
pilkit==2.0
Pillow==9.1.1
python-decouple==3.6
pytz==2022.1
PyYAML==6.0
requests==2.27.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tablib==3.2.1
urllib3==1.26.9
xlrd==2.0.1
xlwt==1.3.0
